+++ src/pages/signup/index.js
@@ -2,7 +2,7 @@ import { postUserData } from '../../apis/userApi';
 import { route } from '../../router/route';
 import { Error503 } from '../../components/common';
 import Form from '../../components/common/form';
-import { FORM_FIELDS, FORM_BUTTONS } from './formFields';
+import { FORM_FIELDS, FORM_BUTTONS } from './formFieldDatas';
 import AvatarImg from '/public/avatar.svg';
 
 export default class SignUpPage {
